+++ CHANGELOG.md
@@ -72,4 +72,5 @@ Read more about this command in the [docs](https://hasura.io/docs/1.0/graphql/ma
 - cli(migrations-docker): add support for v2 config (close #3969)
 
   A new CLI migrations image is introduced to account for the new CLI workflow. If you're have a project with `version: 2` in `config.yaml`, you should use the new image: `hasura/graphql-engine:v1.2.0-cli-migrations-v2`. Mount the migrations at `/hasura-migrations` and metadata at `/hasura-metadata`.
-  
+
+- server: Skip the schema cache building in `run_sql` request, if `skip_cache_reload` is set to `true`. (close #4200)